+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.16.0
  - feat: add `NesIcons.camera`
  - chore: Removing usage from deprecated members from Flutter 3.18.
+ - fix!: `NesDialog`s are now aware of the a `NesFixedViewport` scaling value.
 
 # 0.15.0
  - feat: add `NesIcons.audioMuted`